+++ realm/realm-library/src/androidTestObjectServer/java/io/realm/CredentialsTests.java
@@ -24,6 +24,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
@@ -119,6 +120,7 @@ public class CredentialsTests {
         SyncCredentials creds = SyncCredentials.nickname("foo", false);
         assertEquals(SyncCredentials.IdentityProvider.NICKNAME, creds.getIdentityProvider());
         assertFalse(creds.getUserInfo().isEmpty());
+        assertFalse((Boolean) creds.getUserInfo().get("is_admin"));
     }
 
     @Test
@@ -126,7 +128,7 @@ public class CredentialsTests {
         String[] invalidInput = {null, ""};
         for (String input : invalidInput) {
             try {
-                SyncCredentials.nickname(input);
+                SyncCredentials.nickname(input, false);
                 fail(input + " should have failed");
             } catch (IllegalArgumentException ignored) {
             }